refactor: dissolve lib/src/domain into content-matched folders
Removes lib/src/domain/ and redistributes its files so each filename
matches its content, with no duplicate filenames across the package:

- domain/failures.dart held two concerns -> split:
    result/failures.dart  (Failure hierarchy)
    result/result.dart    (Result/Success/Error/FutureResult/ResultExtension)
- domain/usecase.dart -> usecase/usecase.dart
- domain/safe_call.dart -> network/safe_remote_call.dart
    (renamed: its function is safeRemoteCall; avoids colliding with the
     existing network/safe_call.dart which returns ApiResponse)

safe_remote_call now uses direct relative imports instead of importing
the package barrel. Barrel exports and the failures test import updated.

flutter analyze: 0 issues. flutter test: 295 passed.
