+++ test/domain/failures/failure_test.dart
@@ -1,4 +1,5 @@
-import 'package:flutter_core/src/domain/failures.dart';
+import 'package:flutter_core/src/result/failures.dart';
+import 'package:flutter_core/src/result/result.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
